Refactor currency view out of content view
